+++ src/main/scala/com/gu/acquisition/model/ReferrerAcquisitionData.scala
@@ -4,6 +4,7 @@ import io.circe.{Decoder, Encoder}
 import io.circe.generic.semiauto._
 import ophan.thrift.componentEvent.ComponentType
 import ophan.thrift.event.{AbTest, AcquisitionSource}
+import play.api.libs.json.{Json, Reads, Writes}
 
 /**
   * Model for acquisition data passed by the referrer.
@@ -29,6 +30,11 @@ object ReferrerAcquisitionData {
   import com.gu.acquisition.instances.acquisitionSource._
   import com.gu.acquisition.instances.componentType._
 
-  implicit val referrerAcquisitionDataEncoder: Encoder[ReferrerAcquisitionData] = deriveEncoder[ReferrerAcquisitionData]
   implicit val referrerAcquisitionDataDecoder: Decoder[ReferrerAcquisitionData] = deriveDecoder[ReferrerAcquisitionData]
+
+  implicit val referrerAcquisitionDataEncoder: Encoder[ReferrerAcquisitionData] = deriveEncoder[ReferrerAcquisitionData]
+
+  implicit val referrerAcquisitionDataReads: Reads[ReferrerAcquisitionData] = Json.reads[ReferrerAcquisitionData]
+
+  implicit val referrerAcquisitionDataWrites: Writes[ReferrerAcquisitionData] = Json.writes[ReferrerAcquisitionData]
 }