+++ src/test/scala/com/gu/acquisition/app/AcquisitionSubmissionSimulation.scala
@@ -3,6 +3,7 @@ package com.gu.acquisition.app
 import akka.actor.ActorSystem
 import akka.http.scaladsl.model.Uri
 import akka.stream.{ActorMaterializer, Materializer}
+import com.gu.acquisition.model.{AcquisitionSubmission, OphanIds}
 import com.gu.acquisition.services.OphanService
 import com.typesafe.scalalogging.StrictLogging
 import ophan.thrift.event.{Acquisition, PaymentFrequency, PaymentProvider}
@@ -11,50 +12,40 @@ import scala.concurrent.{ExecutionContext, Future}
 
 trait MockDataGenerator {
 
-  def generateAcquisition: Acquisition
-
-  def generateBrowserId: String
-
-  def generateViewId: String
-
-  def generateVisitId: String
+  def generateSubmission: AcquisitionSubmission
 }
 
 object BasicMockDataGenerator extends MockDataGenerator {
 
-  override def generateAcquisition = Acquisition(
-    product = ophan.thrift.event.Product.Contribution,
-    paymentFrequency = PaymentFrequency.OneOff,
-    currency = "GBP",
-    amount = 20d,
-    amountInGBP = None,
-    paymentProvider = Some(PaymentProvider.Stripe),
-    campaignCode = None,
-    abTests = None,
-    countryCode = Some("US"),
-    referrerPageViewId = None,
-    referrerUrl = None,
-    componentId = None,
-    componentTypeV2 = None,
-    source = None
-  )
-
-  override def generateBrowserId: String = "browserId"
-
-  override def generateViewId: String = "viewId"
-
-  override def generateVisitId: String = "visitId"
+  override def generateSubmission: AcquisitionSubmission =
+    AcquisitionSubmission(
+      OphanIds("pageviewId", Some("visitId"), Some("browserId")),
+      Acquisition(
+        product = ophan.thrift.event.Product.Contribution,
+        paymentFrequency = PaymentFrequency.OneOff,
+        currency = "GBP",
+        amount = 20d,
+        amountInGBP = None,
+        paymentProvider = Some(PaymentProvider.Stripe),
+        campaignCode = None,
+        abTests = None,
+        countryCode = Some("US"),
+        referrerPageViewId = None,
+        referrerUrl = None,
+        componentId = None,
+        componentTypeV2 = None,
+        source = None
+      )
+    )
 }
 
 class AcquisitionSubmissionSimulator(service: OphanService, generator: MockDataGenerator)(implicit ec: ExecutionContext)
   extends StrictLogging {
-
   import cats.instances.future._
-  import generator._
 
   def submit(): Future[Unit] = {
     logger.info("submitting acquisition event to Ophan")
-    service.submit(generateAcquisition, generateViewId, Some(generateBrowserId),  Some(generateVisitId))
+    service.submit(generator.generateSubmission)
       .fold(
         err => logger.error(s"failed to send acquisition to Ophan", err),
         _ => logger.info("successfully submitted acquisition to Ophan")